Edit Team Page
A page that allows users to edit their own team. Admin can edit all teams. All fields are required except for logo image. Additionally, there is a checkbox that allows users to delete the logo image. The logo image uploaded by the user will be resize to 200x200 and 400x400. Finally, the user also have an option to delete the whole team.