Helper functions for Logo
First function will return a string of the path where the logo will be stored. Second function will validate if the file upload is an image. Third function will resize the image using the Gumlet ImageResize package. The last function will delete both the logo and the thumbnail of a team.