Header Update
Leagues required for the header will be stored in "$s_leagues" to prevent changes to "$leagues". New links to new pages added to navigation. Finally a search form is added to allow users to search for teams with the ability to limit the search by league name on every page.