Sign In Page
User authentication system. If the user is already signed in, they are redirected to the Home Page. Connects to a database, validates user input, and checks if the provided username and password match a user in the database. If successful, the user is authenticated and redirected to the Home Page; otherwise, an error message is displayed on the sign-in page. Includes a form for users to input their username and password. Additionally, there is a link to the sign-up page.