Create League Page
Create a new league. The form on the page allows users to input a league name. The code validates the input and, if valid, inserts the new league into the database. Any errors in the input are displayed on the page. After successfully creating the league, the user is redirected to the league details page.